backend/tests/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
open-clip-torch
transformers
pillow
pymilvus
numpy
pydantic
python-multipart